Add basic structure of the project
The code is *wonerful* and *clean* and *well-written*. Truly
astonishing, if I may say so myself. Pure HTML+CSS+JS. Always a
wonderful experience for the ideomatician with a goal. That does sound
almost like magician whose big secret is that he follows the ideomatic
ways of doing things. Is that what he is? I do not yet know. However, we
shall fight to find out the truth of the matter! We shall fight day and
night, in rain and sunshine, in good and in bad. We shall fight until
the day that I shall cease to write wonderful commit messages like this.